Reorganize file tree
Create folders for modals and for logic. Delete components and navigation folders. Collect all files that export logical helpers in /logic.  Rename gameLogic as connectFourLogic.
