+++ src/modals/InfoPanel.js
