+++ src/pages/GameBoard.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 // Logic
-import { gameIsOver, nextPlayerColor } from '../gameLogic'
+import { gameIsOver, nextPlayerColor } from '../logic/connectFourLogic'
 // Custom Hooks
 import { useScreenWidth, useScreenHeight } from "../hooks"
